Release jobs which cause exceptions back to the queue
From @pulkitjalan's patch 5ae35aec11051dbe8dd2a02489c8c39bd0eee0c2

If an exception is encountered while working on a queued job to submit
an error to Sentry, this exception would be caught and added to the
queue as another job. This can snowball and cause all sorts of
problems.

Instead, catch any exception thrown when trying to submit an error and
ignore it, instead releasing the original error submission job back to
the queue after a delay.

Note that this does not protect against similar issues when
synchronously submitting error reports.